add tabs to profile and map
