--- app/src/main/java/com/github/epfl/meili/home/RequiresLoginActivity.kt
@@ -1,30 +1,0 @@
-package com.github.epfl.meili.home
-
-import android.content.Intent
-import android.os.Bundle
-import android.util.Log
-import androidx.appcompat.app.AppCompatActivity
-import com.github.epfl.meili.R
-import com.github.epfl.meili.profile.ProfileActivity
-
-abstract class RequiresLoginActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_requires_login)
-
-        Auth.isLoggedIn.observe(this) {
-            Log.d(ProfileActivity.TAG, "value received $it")
-            verifyAndUpdateUserIsLoggedIn(it)
-        }
-
-        verifyAndUpdateUserIsLoggedIn(Auth.isLoggedIn.value!!)
-    }
-
-    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        super.onActivityResult(requestCode, resultCode, data)
-
-        Auth.onActivityResult(this, requestCode, resultCode, data)
-    }
-
-    abstract fun verifyAndUpdateUserIsLoggedIn(isLoggedIn: Boolean)
-}
