+++ app/src/main/res/layout/activity_main.xml
@@ -3,7 +3,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     tools:context=".MainActivity">
 
     <Button
@@ -77,5 +76,4 @@
         android:layout_below="@id/launchPoiView"
         android:text="Launch Nearby"/>
 
-    <include layout="@layout/activity_navigable"/>
 </RelativeLayout>
